feat: optimize
