--- src/assets/scss/base/_functions.scss
@@ -1,1 +1,0 @@
-// Function styles goes here
