+++ src/assets/scss/base/_global.scss
@@ -4,9 +4,6 @@ body {
   height: 100%;
   width: 100%;
 }
-// body {
-//   overflow: hidden;
-// }
 
 #app {
   height: inherit;
