+++ src/assets/scss/base/_index.scss
@@ -3,4 +3,3 @@
 @import "mixins";
 @import "placeholder";
 @import "global";
-@import "functions";
