+++ src/assets/scss/base/_placeholder.scss
@@ -4,3 +4,8 @@
   background-size: cover;
   background-repeat: no-repeat;
 }
+%flex-middle {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
