+++ src/assets/scss/base/_variables.scss
@@ -1,5 +1,9 @@
+// Font
 $font-family-base: "Poppins", sans-serif;
 $font-family-cursive: "Courgette", cursive;
 
 $font-size-base: 1rem;
 $font-size-headings: $font-size-base * 4;
+
+// transition
+$primary-transition: all 0.35s ease-in-out;
