+++ src/assets/scss/component/_breadcrumb.scss
@@ -11,15 +11,12 @@
   }
   .cart_item_count {
     font-size: 11px;
-    display: inline-block;
     padding: 3px;
     height: 20px;
     width: 20px;
     // top: -8px;
     // right: -8px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
+    @extend %flex-middle;
     font-family: $font-family-base;
   }
 }
