+++ src/assets/scss/layout/_about-hero.scss
@@ -17,7 +17,7 @@
     h2 {
       font-size: 40px;
     }
-    h3 {
+    p {
       font-size: 15px !important;
       font-weight: 400 !important;
     }
