+++ src/assets/scss/layout/_hero-section.scss
@@ -1,3 +1,4 @@
+// Home Section
 .hero-section {
   position: relative;
   background-image: url("~@/assets/images/heroSection/heroDesktopMode.jpg");
@@ -8,7 +9,7 @@
   @media (max-width: 576px) {
     background-attachment: unset;
   }
-
+  // decrease the visibility of background-image
   .hero_overlay {
     position: absolute;
     background-color: black;
@@ -19,6 +20,7 @@
     bottom: 0;
     z-index: 1;
   }
+
   .hero_content {
     z-index: 10;
     & > span {
@@ -52,6 +54,7 @@
   @media (max-width: 576px) {
     height: 500px;
   }
+  // hero button
   .order-button {
     @media (max-width: 350px) {
       padding: 10px !important;
