+++ src/assets/scss/layout/_index.scss
@@ -8,6 +8,6 @@
 @import "product-details";
 @import "contact-us";
 @import "about-hero";
-@import "./page-not-found";
-@import "about";
+@import "about-feed";
 @import "footer";
+@import "./page-not-found";
