+++ src/assets/scss/layout/_product-card.scss
@@ -1,3 +1,6 @@
+.product_card {
+  cursor: pointer;
+}
 .product_image_wrapper {
   height: 250px;
   overflow: hidden;
