--- src/component/layout/OrderHero.vue
@@ -1,11 +1,0 @@
-<template>
-  <section class="order_hero_wrapper position-relative">
-    <div class="overlay"></div>
-    <div
-      class="hero_data text-white position-absolute top-50 start-50 translate-middle text-center"
-    >
-      <h2>Order Online</h2>
-      <p class="hero-para fs-6">What's on your mind?</p>
-    </div>
-  </section>
-</template>
