+++ src/component/order/BreakfastList.vue
@@ -1,0 +1,25 @@
+<template>
+  <section class="">
+    <h3 class="fs-3">Breakfast</h3>
+    <div class="row mt-3 gx-5 gy-5">
+      <BaseProductItem
+        v-for="(breakfast, index) in allBreakfastProduct"
+        :key="index"
+        :product="breakfast"
+        @getDetails="$emit('getDetails', $event)"
+      ></BaseProductItem>
+    </div>
+  </section>
+</template>
+<script lang="ts">
+import { inject, defineComponent } from "vue";
+import BaseProductItem from "@/ui/BaseProductItem.vue";
+export default defineComponent({
+  components: { BaseProductItem },
+  //   props: ["allBreakfastProduct"],
+  setup() {
+    const allBreakfastProduct: any = inject("allBreakfastProduct");
+    return { allBreakfastProduct: allBreakfastProduct };
+  },
+});
+</script>
