--- src/component/product/bakery/BakeryItem.vue
@@ -1,28 +1,0 @@
-<template>
-  <div class="col-lg-4 col-sm-6 col-12">
-    <div
-      class="border product_card shadow card p-3"
-      @click="$emit('getDetails', bakery)"
-    >
-      <div class="product_image_wrapper rounded">
-        <img :src="bakery.imageUrl" alt="" />
-      </div>
-      <div class="mt-3">
-        <h5 class="fs-5">{{ bakery.title }}</h5>
-        <p class="">
-          {{
-            bakery.description.length > 50
-              ? bakery.description.substring(0, 50 - 3) + "..."
-              : bakery.description
-          }}
-        </p>
-        <span>${{ bakery.price }}</span>
-      </div>
-    </div>
-  </div>
-</template>
-<script lang="ts">
-export default {
-  props: ["bakery"],
-};
-</script>
