--- src/component/product/bakery/BakeryList.vue
@@ -1,25 +1,0 @@
-<template>
-  <section>
-    <h3 class="fs-3">Bakery</h3>
-    <div class="row mt-3 gx-5 gy-5">
-      <BakeryItem
-        v-for="(bakery, index) in allBakeryProduct"
-        :key="index"
-        :bakery="bakery"
-        @getDetails="$emit('getDetails', $event)"
-      ></BakeryItem>
-    </div>
-  </section>
-</template>
-<script lang="ts">
-import BakeryItem from "./BakeryItem.vue";
-import { inject } from "vue";
-export default {
-  // props: ["allBakeryProduct"],
-  components: { BakeryItem },
-  setup() {
-    const allBakeryProduct: any = inject("allBakeryProduct");
-    return { allBakeryProduct: allBakeryProduct };
-  },
-};
-</script>
