+++ src/store/auth/getters.ts
@@ -13,5 +13,8 @@ export default ({
     },
     showGuardAlert(state: any) {
         return state.showGuardAlert
+    },
+    token(state: any) {
+        return state.token
     }
 })
