+++ src/views/ContactView.vue
@@ -1,11 +1,14 @@
 <template>
   <section class="contact_us theme-p mx-auto d-flex align-items-center h-100">
+    <!-- Submit Form Error Message Dialog -->
     <base-dialog :show="!!catchError" @close="handleError" :title="catchError">
       <p>something went wrong please try after sometimes</p>
     </base-dialog>
     <base-card class="w-100" v-if="!submitted">
       <div class="d-flex flex-column align-items-center justify-content-center">
+        <!-- heading  -->
         <h1>Contact Us</h1>
+        <!-- Form Section  -->
         <div class="w-100">
           <Form
             @submit="onSubmit"
@@ -13,8 +16,11 @@
             v-slot="{ errors }"
           >
             <div class="row mt-3">
+              <!-- Firstname  -->
               <div class="col-6">
-                <label for="firstName">First Name</label>
+                <label for="firstName"
+                  >Firstname<span class="text-danger">*</span></label
+                >
                 <Field
                   name="firstName"
                   id="firstName"
@@ -23,15 +29,18 @@
                   v-model="activeUser.firstName"
                   :class="{ 'is-invalid': errors.firstName }"
                 />
-                <!-- :value="activeUser?.firstName || ''" -->
+                <!-- Firstname validation error message  -->
                 <span
                   class="invalid-text text-danger"
                   v-if="errors.firstName"
                   >{{ errors.firstName }}</span
                 >
               </div>
+              <!-- Lastname  -->
               <div class="col-6">
-                <label for="lastName">Last Name</label>
+                <label for="lastName"
+                  >Lastname<span class="text-danger">*</span></label
+                >
                 <Field
                   name="lastName"
                   id="lastName"
@@ -40,14 +49,15 @@
                   v-model="activeUser.lastName"
                   :class="{ 'is-invalid': errors.lastName }"
                 />
-                <!-- :value="activeUser?.lastName || ''" -->
+                <!-- Lastname validation error message  -->
                 <span class="invalid-text text-danger" v-if="errors.lastName">{{
                   errors.lastName
                 }}</span>
               </div>
             </div>
+            <!-- Email  -->
             <div class="mt-3">
-              <label for="email">Email</label>
+              <label for="email">Email<span class="text-danger">*</span></label>
               <Field
                 name="email"
                 id="email"
@@ -55,17 +65,20 @@
                 placeholder="e.g. test@gmail.com"
                 :class="{ 'is-invalid': errors.email }"
                 v-model="activeUser.email"
+                autocomplete=""
               />
-              <!-- :value="activeUser?.email || ''" -->
+              <!-- email validation error message  -->
               <span class="invalid-text text-danger" v-if="errors.email">{{
                 errors.email
               }}</span>
             </div>
+            <!-- Message area  -->
             <div class="mt-3">
-              <label for="" class="my-2"
-                >Message <span class="text-danger">*</span>
+              <label for="message" class="my-2"
+                >Message<span class="text-danger">*</span>
               </label>
               <Field
+                id="message"
                 as="textarea"
                 name="message"
                 class="form-control"
@@ -74,10 +87,12 @@
                 placeholder="Enter your message here"
                 :class="{ 'is-invalid': errors.message }"
               ></Field>
+              <!-- message validation error message  -->
               <span class="invalid-text text-danger" v-if="errors.message">{{
                 errors.message
               }}</span>
             </div>
+            <!-- For Submit and back button  -->
             <div class="mt-3">
               <button type="submit" class="btn btn-primary me-2 text-nowrap">
                 Submit
@@ -91,9 +106,11 @@
               </button>
             </div>
           </Form>
+          <!-- End  -->
         </div>
       </div>
     </base-card>
+    <!-- To showing success Message When form is submitted  -->
     <base-card class="w-100" v-else>
       <div class="">
         Form Submitted Successfully Go back to
@@ -104,6 +121,7 @@
 </template>
 
 <script lang="ts">
+// Imports
 import * as yup from "yup";
 import { Field, Form, FormContext } from "vee-validate";
 import { defineComponent, ref, computed, watch } from "vue";
@@ -139,7 +157,7 @@ export default defineComponent({
         if (currentUser.value) {
           activeUser.value = currentUser.value;
         }
-        console.log(currentUser.value);
+        // console.log(currentUser.value);
       },
       { immediate: true }
     );
@@ -166,7 +184,7 @@ export default defineComponent({
       }
       const time = getHours + ":" + getMinutes;
       const day = getFullDay;
-      console.log(day);
+      // console.log(day);
 
       // Sending contact request
       try {
